SF36

Documentation for SF36.

SF36.sf12

SF36.sf36

SF36.sf12 — Method

sf12(v::AbstractVector{Int}; version = :v1)

Calculate SF12 results for vector of answers v.

1 - GH1 2 - PF02 3 - PF04 4 - RP2 5 - RP3 6 - RE2 7 - RE3 8 - BP2 9 - MH3 10 - VT2 11 - MH4 12 - SF2

Reference:

Ware Jr, John E., Mark Kosinski, and Susan D. Keller. "A 12-Item Short-Form Health Survey: construction of scales and preliminary tests of reliability and validity." Medical care 34.3 (1996): 220-233.Validation Literature:

Gandek, Barbara, et al. "Cross-validation of item selection and scoring for the SF-12 Health Survey in nine countries: results from the IQOLA Project." Journal of clinical epidemiology51.11 (1998): 1171-1178. Jenkinson, Crispin, et al. "A shorter form health survey: can the SF-12 replicate results from the SF-36 in longitudinal studies?." Journal of Public Health 19.2 (1997): 179-186.Additional Literature:

Ware, John E., Susan D. Keller, and Mark Kosinski. SF-12: How to score the SF-12 physical and mental health summary scales. Health Institute, New England Medical Center, 1995.

See also: https://orthotoolkit.com/sf-12/

https://www.researchgate.net/profile/John-Ware-6/publication/291994160HowtoscoreSF-12_items/links/58dfc42f92851c369548e04e/How-to-score-SF-12-items.pdf

SF36.sf36 — Method

sf36(v::AbstractVector; version = :v1)

Calculate SF36 results for vector of answers v.

version keyword:

:v1 - General US population (Ware, John & MA, Kosinski & Keller, S.D.. (1993). SF-36 Physical and Mental Health Summary Scales: a User's Manual. 8. 23-28.)

:v2 - Correlated scores (Grassi, M. and Nucera, A. (2010), Dimensionality and Summary Measures of the SF-36 v1.6: Comparison of Scale- and Item-Based Approach Across ECRHS II Adults Population. Value in Health, 13: 469-478.)

No other available yet.

Vector v:

1,33:36 - General Health(GH) - 1, 11a:11d 3:12 - Physical functioning (PF) - 3a:3j 13:16 - Role-Physical Functioning (RP) - 4a:4d 17:19 - RoleEmotional (RE) - 5a:5c 20,32 - Social Functioning (SF) - 6,10 21,22 - Bodily pain (BP) - 7,8 24:26,28,30 - Mental Health (MH) - 9b:9d, 9f, 9h 23,27,29,31 - Vitality (VT) - 9a, 9e, 9g, 9i