SF36HealthSurvey

Documentation for SF36HealthSurvey.

The Short Form (36) Health Survey is a 36-item, patient-reported survey of patient health. The SF-36 is a measure of health status and an abbreviated variant of it, the SF-6D, is commonly used in health economics as a variable in the quality-adjusted life year calculation to determine the cost-effectiveness of a health treatment. The original SF-36 stemmed from the Medical Outcome Study, MOS, which was conducted by the RAND Corporation. SF36 Version 1 - is a free instrument.

Terms and Conditions for Using the 36-Item Short Form Survey (SF-36):

RAND hereby grants permission to use RAND 36-Item Short Form Health Survey in accordance with the following conditions, which shall be assumed by all to have been agreed to as a consequence of accepting and using this document:

Changes to the Health Survey may be made without the written permission of RAND. However, all such changes shall be clearly identified as having been made by the recipient.
The user of this Health Survey accepts full responsibility, and agrees to indemnify and hold RAND harmless, for the accuracy of any translations of the Health Survey into another language and for any errors, omissions, misinterpretations, or consequences thereof.
The user of this Health Survey accepts full responsibility, and agrees to indemnify and hold RAND harmless, for any consequences resulting from the use of the Health Survey.
The user of the 36-Item Health Survey will provide a credit line when printing and distributing this document acknowledging that it was developed at RAND as part of the Medical Outcomes Study.
No further written permission is needed for use of this Health Survey.

SF36HealthSurvey.sf12

SF36HealthSurvey.sf36

SF36HealthSurvey.sf12 — Method

sf12(v::AbstractVector{Int}; version = :v1)

Calculate SF12 results for vector of answers v.

1 - GH1 2 - PF02 3 - PF04 4 - RP2 5 - RP3 6 - RE2 7 - RE3 8 - BP2 9 - MH3 10 - VT2 11 - MH4 12 - SF2

Reference:

Ware Jr, John E., Mark Kosinski, and Susan D. Keller. "A 12-Item Short-Form Health Survey: construction of scales and preliminary tests of reliability and validity." Medical care 34.3 (1996): 220-233.Validation Literature:

Gandek, Barbara, et al. "Cross-validation of item selection and scoring for the SF-12 Health Survey in nine countries: results from the IQOLA Project." Journal of clinical epidemiology51.11 (1998): 1171-1178. Jenkinson, Crispin, et al. "A shorter form health survey: can the SF-12 replicate results from the SF-36 in longitudinal studies?." Journal of Public Health 19.2 (1997): 179-186.Additional Literature:

Ware, John E., Susan D. Keller, and Mark Kosinski. SF-12: How to score the SF-12 physical and mental health summary scales. Health Institute, New England Medical Center, 1995.

See also: https://orthotoolkit.com/sf-12/

https://www.researchgate.net/profile/John-Ware-6/publication/291994160HowtoscoreSF-12_items/links/58dfc42f92851c369548e04e/How-to-score-SF-12-items.pdf

SF36HealthSurvey.sf36 — Method

sf36(v::AbstractVector; version = :v1)

Calculate SF36 results for vector of answers v.

version keyword:

:v1 - General US population (Ware, John & MA, Kosinski & Keller, S.D.. (1993). SF-36 Physical and Mental Health Summary Scales: a User's Manual. 8. 23-28.)

:v2 - Correlated scores (Grassi, M. and Nucera, A. (2010), Dimensionality and Summary Measures of the SF-36 v1.6: Comparison of Scale- and Item-Based Approach Across ECRHS II Adults Population. Value in Health, 13: 469-478.)

No other available yet.

Vector v:

1,33:36 - General Health(GH) - 1, 11a:11d 3:12 - Physical functioning (PF) - 3a:3j 13:16 - Role-Physical Functioning (RP) - 4a:4d 17:19 - RoleEmotional (RE) - 5a:5c 20,32 - Social Functioning (SF) - 6,10 21,22 - Bodily pain (BP) - 7,8 24:26,28,30 - Mental Health (MH) - 9b:9d, 9f, 9h 23,27,29,31 - Vitality (VT) - 9a, 9e, 9g, 9i